Move PanoramaDemo to ExampleBrowser
