+++ Source/Examples/WPF/ExampleBrowser/Examples/PanoramaDemo/Models/GrandHotel/cube_b.JPG
