+++ Source/Examples/WPF/ExampleBrowser/Examples/PanoramaDemo/Models/Opera/cube_b.JPG
