+++ Source/Examples/WPF/ExampleBrowser/Examples/PanoramaDemo/Models/Opera/cube_l.JPG
