+++ Source/HelixToolkit.Wpf.sln
@@ -15,8 +15,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HelixToolkit.Wpf.Input", "H
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HelixToolkit.Wpf", "HelixToolkit.Wpf\HelixToolkit.Wpf.csproj", "{3F3B4548-A9DB-4874-86F7-E7C4FEC8A351}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PanoramaDemo", "Examples\WPF\PanoramaDemo\PanoramaDemo.csproj", "{E68C0BD5-316A-4546-9FFC-9722350E4467}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tools", "Tools", "{4EC16C60-5AFA-4854-82D9-FB1BF92C3812}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BatchRender", "Tools\BatchRender\BatchRender.csproj", "{0DA2D747-B3FB-40A6-B8FF-0F8949A3E1BC}"
@@ -95,17 +93,6 @@ Global
 		{3F3B4548-A9DB-4874-86F7-E7C4FEC8A351}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
 		{3F3B4548-A9DB-4874-86F7-E7C4FEC8A351}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{3F3B4548-A9DB-4874-86F7-E7C4FEC8A351}.Release|x86.ActiveCfg = Release|Any CPU
-		{E68C0BD5-316A-4546-9FFC-9722350E4467}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{E68C0BD5-316A-4546-9FFC-9722350E4467}.Debug|Any CPU.Build.0 = Debug|x86
-		{E68C0BD5-316A-4546-9FFC-9722350E4467}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
-		{E68C0BD5-316A-4546-9FFC-9722350E4467}.Debug|Mixed Platforms.Build.0 = Debug|x86
-		{E68C0BD5-316A-4546-9FFC-9722350E4467}.Debug|x86.ActiveCfg = Debug|x86
-		{E68C0BD5-316A-4546-9FFC-9722350E4467}.Debug|x86.Build.0 = Debug|x86
-		{E68C0BD5-316A-4546-9FFC-9722350E4467}.Release|Any CPU.ActiveCfg = Release|x86
-		{E68C0BD5-316A-4546-9FFC-9722350E4467}.Release|Mixed Platforms.ActiveCfg = Release|x86
-		{E68C0BD5-316A-4546-9FFC-9722350E4467}.Release|Mixed Platforms.Build.0 = Release|x86
-		{E68C0BD5-316A-4546-9FFC-9722350E4467}.Release|x86.ActiveCfg = Release|x86
-		{E68C0BD5-316A-4546-9FFC-9722350E4467}.Release|x86.Build.0 = Release|x86
 		{0DA2D747-B3FB-40A6-B8FF-0F8949A3E1BC}.Debug|Any CPU.ActiveCfg = Debug|x86
 		{0DA2D747-B3FB-40A6-B8FF-0F8949A3E1BC}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
 		{0DA2D747-B3FB-40A6-B8FF-0F8949A3E1BC}.Debug|Mixed Platforms.Build.0 = Debug|x86
@@ -184,7 +171,6 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{B262B3C4-B4F7-4CD0-9055-62B271856078} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
 		{CD249FE2-999C-4AC3-810D-C53ACB2A7DF3} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
-		{E68C0BD5-316A-4546-9FFC-9722350E4467} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
 		{0DA2D747-B3FB-40A6-B8FF-0F8949A3E1BC} = {4EC16C60-5AFA-4854-82D9-FB1BF92C3812}
 		{47FC8A4C-86B6-4D9D-BE10-EE74F8B992BB} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
 		{991A4C2B-D830-41CB-A810-047ADCF69219} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
